Fix compiler warning with using size_t as unsigned long.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/gpl-compat@26377 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/FileSystem/DAVDirectory.cpp
@@ -71,7 +71,7 @@ bool CDAVDirectory::ValueWithoutNamespace(const TiXmlNode *pNode, CStdString val
   }
   else if (result.size() > 2)
   {
-    CLog::Log(LOGERROR, "%s - Splitting %s failed, size(): %lu, value: %s", __FUNCTION__, pElement->Value(), result.size(), value.c_str());
+    CLog::Log(LOGERROR, "%s - Splitting %s failed, size(): %lu, value: %s", __FUNCTION__, pElement->Value(), (unsigned long)result.size(), value.c_str());
   }
     
   return false;
